+++ GeoIP2/Model/Continent.cs
@@ -4,9 +4,7 @@ using Newtonsoft.Json;
 namespace MaxMind.GeoIP2.Model
 {
     /// <summary>
-    /// Contains data for the continent record associated with an IP address.
-    ///  
-    ///  This record is returned by all the end points.    
+    /// Contains data for the continent record associated with an IP address.  
     /// </summary>
     public class Continent : NamedEntity
     {
@@ -26,7 +24,7 @@ namespace MaxMind.GeoIP2.Model
 
         /// <summary>
         /// A two character continent code like "NA" (North America) or "OC"
-        /// (Oceania). This attribute is returned by all end points.        
+        /// (Oceania).      
         /// </summary>
         [JsonProperty("code")]
         public string Code { get; internal set; }