+++ GeoIP2/Responses/InsightsResponse.cs
@@ -6,7 +6,7 @@ namespace MaxMind.GeoIP2.Responses
 {
     /// <summary>
     /// This class provides a model for the data returned by the GeoIP2 Precision: 
-    /// Insights web service end point.
+    /// Insights end point.
     /// 
     /// The only difference between the City and Insights response classes is
     /// which fields in each record may be populated.
@@ -14,18 +14,17 @@ namespace MaxMind.GeoIP2.Responses
     /// <a href="http://dev.maxmind.com/geoip/geoip2/web-services">GeoIP2 Web
     ///      Services</a>
     /// </summary>
-    [Obsolete("OmniResponse is deprecated. Please use InsightsResponse instead.")]
-    public class OmniResponse : AbstractCityResponse
+    public class InsightsResponse : AbstractCityResponse
     {
         /// <summary>
         /// Constructor
         /// </summary>
-        public OmniResponse() { }
+        public InsightsResponse() { }
 
         /// <summary>
         /// Constructor
         /// </summary>
-        public OmniResponse(
+        public InsightsResponse(
             City city = null,
             Continent continent = null,
             Country country = null,